fix-no-newline-parse (#102)
* fix-no-newline-parse

Dearest Reviewer,

Issue #90 brought up the issue that if there was no newline. the user
provided an easy reproduction case. I was able to track it down to the
parse raw function.

The issue is that the loop hits the end of the string and returns Empty.
However, if there is a start offset then some characters were found. This
code just checks the offset after the loop.

If this is not the correct behavior please close.

Thanks for your time
Becker

* pr feedback

Add break instead of duplicating code.