refactor: always return groups when parsing address lists
The enum separating the Vec<Addr> from the Vec<Group> case carries a large
API burden with it.

The Address::List variant introduces code duplication to any consumer which has to write
code to handle both Address::List and Address::Group, even though much of this code
will be the same due to the way RFC8621 4.1.2.4 GroupedAddresses deals with single mailboxes.

By reducing the enum variants, this duplication is no longer necessary.